Fix namespaces and move visual tests to graphics
