+++ Test/Framework/Graphics/GraphicsDeviceManagerTest.cs
@@ -7,7 +7,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     internal class GraphicsDeviceManagerTest
