+++ Test/Framework/Graphics/GraphicsDeviceTest.cs
@@ -7,10 +7,9 @@ using System.Linq;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Tests.ContentPipeline;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     internal class GraphicsDeviceTest : GraphicsDeviceTestFixtureBase
