+++ Test/Framework/Graphics/IndexBufferTest.cs
@@ -2,7 +2,6 @@
 // This file is subject to the terms and conditions defined in
 // file 'LICENSE.txt', which is part of this source code package.
 
-using System;
 using NUnit.Framework;
 using Microsoft.Xna.Framework.Graphics;
