+++ Test/Framework/Graphics/OcclusionQueryTest.cs
@@ -5,14 +5,39 @@
 using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     internal class OcclusionQueryTest : GraphicsDeviceTestFixtureBase
     {
+
+        [Test]
+        public void ConstructorsAndProperties()
+        {
+            Assert.Throws<ArgumentNullException>(() => new OcclusionQuery(null));
+
+            var occlusionQuery = new OcclusionQuery(gd);
+
+            Assert.IsFalse(occlusionQuery.IsComplete);
+
+            Assert.Throws<InvalidOperationException>(
+                () => { var n = occlusionQuery.PixelCount; },
+                "PixelCount throws when query not yet started.");
+        }
+
+        [Test]
+        public void MismatchedBeginEnd()
+        {
+            var occlusionQuery = new OcclusionQuery(gd);
+
+            Assert.Throws<InvalidOperationException>(() => occlusionQuery.End());
+
+            occlusionQuery.Begin();
+            Assert.Throws<InvalidOperationException>(() => occlusionQuery.Begin());
+        }
+
         [Test]
         public void QueryOccludedSprites()
         {
