+++ Test/Framework/Graphics/RasterizerStateTest.cs
@@ -2,17 +2,61 @@
 // This file is subject to the terms and conditions defined in
 // file 'LICENSE.txt', which is part of this source code package.
 
+using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Tests.Components;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     internal class RasterizerStateTest : GraphicsDeviceTestFixtureBase
     {
+        [Test]
+        public void ShouldNotBeAbleToSetNullRasterizerState()
+        {
+            Assert.Throws<ArgumentNullException>(() => gd.RasterizerState = null);
+        }
+
+        [Test]
+        public void ShouldNotBeAbleToMutateStateObjectAfterBindingToGraphicsDevice()
+        {
+            var rasterizerState = new RasterizerState();
+
+            // Can mutate before binding.
+            DoAsserts(rasterizerState, Assert.DoesNotThrow);
+
+            // Can't mutate after binding.
+            gd.RasterizerState = rasterizerState;
+            DoAsserts(rasterizerState, d => Assert.Throws<InvalidOperationException>(d));
+
+            // Even after changing to different RasterizerState, you still can't mutate a previously-bound object.
+            gd.RasterizerState = RasterizerState.CullCounterClockwise;
+            DoAsserts(rasterizerState, d => Assert.Throws<InvalidOperationException>(d));
+        }
+
+        [Test]
+        public void ShouldNotBeAbleToMutateDefaultStateObjects()
+        {
+            DoAsserts(RasterizerState.CullClockwise, d => Assert.Throws<InvalidOperationException>(d));
+            DoAsserts(RasterizerState.CullCounterClockwise, d => Assert.Throws<InvalidOperationException>(d));
+            DoAsserts(RasterizerState.CullNone, d => Assert.Throws<InvalidOperationException>(d));
+        }
+
+        private static void DoAsserts(RasterizerState rasterizerState, Action<TestDelegate> assertMethod)
+        {
+            assertMethod(() => rasterizerState.CullMode = CullMode.CullClockwiseFace);
+            assertMethod(() => rasterizerState.DepthBias = 0.1f);
+            assertMethod(() => rasterizerState.FillMode = FillMode.WireFrame);
+            assertMethod(() => rasterizerState.MultiSampleAntiAlias = true);
+            assertMethod(() => rasterizerState.ScissorTestEnable = true);
+            assertMethod(() => rasterizerState.SlopeScaleDepthBias = 0.2f);
+#if !XNA
+            assertMethod(() => rasterizerState.DepthClipEnable = false);
+#endif
+        }
+
         [TestCase(CullMode.CullClockwiseFace)]
         [TestCase(CullMode.CullCounterClockwiseFace)]
         [TestCase(CullMode.None)]
