+++ Test/Framework/Graphics/RenderTargetCubeTest.cs
@@ -4,10 +4,9 @@
 
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     class RenderTargetCubeTest : GraphicsDeviceTestFixtureBase
