+++ Test/Framework/Graphics/SamplerStateTest.cs
@@ -6,10 +6,9 @@ using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Tests.ContentPipeline;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     internal class SamplerStateTest : GraphicsDeviceTestFixtureBase
