+++ Test/Framework/Graphics/Texture2DTest.cs
@@ -70,10 +70,9 @@ using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Tests.ContentPipeline;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     class Texture2DTest : GraphicsDeviceTestFixtureBase
