+++ Test/Framework/Graphics/Texture3DNonVisualTest.cs
@@ -3,14 +3,9 @@
 // file 'LICENSE.txt', which is part of this source code package.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using NUnit.Framework;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using NUnit.Framework.Constraints;
 
 namespace MonoGame.Tests.Graphics
 {
