+++ Test/Framework/Graphics/TextureCubeTest.cs
@@ -4,10 +4,9 @@
 
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using MonoGame.Tests.Graphics;
 using NUnit.Framework;
 
-namespace MonoGame.Tests.Visual
+namespace MonoGame.Tests.Graphics
 {
     [TestFixture]
     class TextureCubeTest : GraphicsDeviceTestFixtureBase
