+++ Test/MonoGame.Tests.XNA.csproj
@@ -120,15 +120,11 @@
     <Compile Include="Framework\GameComponentTest.cs" />
     <Compile Include="Framework\GamePadStateTest.cs" />
     <Compile Include="Framework\GamePadThumbSticksTest.cs" />
-    <Compile Include="Framework\Graphics\DepthStencilStateNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\EffectTest.cs" />
     <Compile Include="Framework\Graphics\GraphicsAdapterTest.cs" />
     <Compile Include="Framework\Graphics\GraphicsDeviceTest.cs" />
     <Compile Include="Framework\Graphics\GraphicsDeviceTestFixtureBase.cs" />
     <Compile Include="Framework\Graphics\MiscellaneousTests.cs" />
-    <Compile Include="Framework\Graphics\OcclusionQueryNonVisualTest.cs" />
-    <Compile Include="Framework\Graphics\RasterizerStateNonVisualTest.cs" />
-    <Compile Include="Framework\Graphics\SpriteFontNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\Texture2DNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\Texture3DNonVisualTest.cs" />
     <Compile Include="Framework\MatrixTest.cs" />
@@ -142,18 +138,7 @@
     <Compile Include="Framework\Vector4Test.cs" />
     <Compile Include="Framework\VertexTests.cs" />
     <Compile Include="Framework\Graphics\BlendStateTest.cs" />
-    <Compile Include="Framework\Visual\DepthStencilStateTest.cs" />
-    <Compile Include="Framework\Visual\GraphicsDeviceManagerTest.cs" />
     <Compile Include="Framework\Graphics\IndexBufferTest.cs" />
-    <Compile Include="Framework\Visual\RasterizerStateTest.cs" />
-    <Compile Include="Framework\Visual\RenderTarget2DTest.cs" />
-    <Compile Include="Framework\Visual\RenderTargetCubeTest.cs" />
-    <Compile Include="Framework\Visual\SamplerStateTest.cs" />
-    <Compile Include="Framework\Visual\TextureCubeTest.cs" />
-    <Compile Include="Framework\Visual\VertexBufferTest.cs" />
-    <Compile Include="Framework\Visual\ScissorRectangleTest.cs" />
-    <Compile Include="Framework\Visual\Texture2DTest.cs" />
-    <Compile Include="Framework\Visual\Texture3DTest.cs" />
     <Compile Include="Runner\Constants.cs" />
     <Compile Include="Runner\Desktop\FramePixelData.cs" />
     <Compile Include="Runner\Extensions.cs" />
@@ -177,15 +162,10 @@
     <Compile Include="Framework\TestGameBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Framework\Components\PixelDeltaFrameComparer.cs" />
-    <Compile Include="Framework\Visual\SpriteBatchTest.cs" />
-    <Compile Include="Framework\Visual\SpriteFontTest.cs" />
     <Compile Include="Framework\Components\UpdateGuard.cs" />
-    <Compile Include="Framework\Visual\ViewportTest.cs" />
-    <Compile Include="Framework\Visual\OcclusionQueryTest.cs" />
     <Compile Include="Framework\Visual\VisualTestFixtureBase.cs" />
     <Compile Include="Framework\Visual\VisualTestGame.cs" />
     <Compile Include="Runner\Windows\Program.cs" />
-    <Compile Include="Framework\Visual\ShaderTest.cs" />
     <Compile Include="Framework\Components\Colored3DCubeComponent.cs" />
     <Compile Include="Framework\Components\TexturedQuadComponent.cs" />
     <Compile Include="Framework\Components\SpaceshipModelDrawComponent.cs" />
